Renamed a package in the org.eclipse.smarthome.io.rest.log bundle. See openhab/static-code-analysis#100
