+++ bundles/io/org.eclipse.smarthome.io.rest.log/src/main/java/org/eclipse/smarthome/io/rest/log/internal/LogConstants.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.io.rest.log;
+package org.eclipse.smarthome.io.rest.log.internal;
 
 /**
  * The {@link LogConstants} class defines common constants, which are
